ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ

+++ app/src/main/jni/giflossy/src/clp.c
@@ -41,7 +41,7 @@
 # include <inttypes.h>
 #endif
 #if !defined(HAVE_UINTPTR_T) && defined(HAVE_CONFIG_H)
-typedef unsigned long uintptr_t;
+typedef unsigned int  uintptr_t;
 #endif
 
 #ifdef __cplusplus
